Recipe fragment is now accessed through home fragment, changes were deleted from main activity
