+++ app/src/main/res/layout/activity_main.xml
@@ -20,10 +20,5 @@
         android:layout_margin="16dp"
         android:src="@drawable/ic_add" />
 
-    <FrameLayout
-        android:id="@+id/recipePlaceholder"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"></FrameLayout>
-
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
