+++ app/src/main/res/layout/fragment_home.xml
@@ -28,5 +28,13 @@
         android:layout_height="match_parent"
         android:src="@drawable/image_waiting" />
 
+    <FrameLayout
+        android:id="@+id/recipePlaceholder"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+    </FrameLayout>
+
+
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
